bad_word_detector/requirements.txt
# API
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# ML & Data Science
torch==2.0.1
transformers==4.33.2
datasets==2.14.5
scikit-learn==1.3.0
nltk==3.8.1
pandas==2.1.0
numpy==1.25.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.1
pytest==7.4.2
tqdm==4.66.1
psutil==5.9.5

# Multilingual Support
langdetect==1.0.9